--- guides/bug_report_templates/active_job_gem.rb
@@ -1,29 +1,0 @@
-# frozen_string_literal: true
-
-require "bundler/inline"
-
-gemfile(true) do
-  source "https://rubygems.org"
-
-  git_source(:github) { |repo| "https://github.com/#{repo}.git" }
-
-  # Activate the gem you are reporting the issue against.
-  gem "activejob", "~> 7.1.0"
-end
-
-require "minitest/autorun"
-require "active_job"
-
-class BuggyJob < ActiveJob::Base
-  def perform
-    puts "performed"
-  end
-end
-
-class BuggyJobTest < ActiveJob::TestCase
-  def test_stuff
-    assert_enqueued_with(job: BuggyJob) do
-      BuggyJob.perform_later
-    end
-  end
-end
